Commit - 4
Built a Validation method for the form

Made use of useRef Hook instead of usestate and handle event.

wrote a validation check function that takes in email and password parameter.

with useRef took email and password from the form , which was stored in current.value

used state variable to store the error message and render the message only if the error message is not null .

Also finished The home work by adding 2 different validation checks for signup and sing in data.